Added the handling of group related error codes in the GroupController. It is created in a way that most common errors, such as GroupNotFound, GroupParticipantNotFound, ConversationNotFound, BallotNotFound, ConversationIsClosed und BallotIsClosed can be handled by a central method that performs standard actions for each of this errors. Specific error handling, if desired, can be performed at each controller method.
Fixed Bug: Trying to store a conversation message caused inconistent datasets when the corresponding conversations has not already been added to the datasets. Added a check in StoreConversationMessage method of GroupController. Even more important, changed the corresponding Database methods to prevent this error. This was done for channels in the ChannelDatabaseManger as well.

Added method MarkBallotAsClosed in the GroupController.

Changed the behavior of handleConversationMessageNewPushMsgAsync in PushNotificationController. In cases the StoreConversationMessage method fails due to missing references, it now performs a complete synchronization of all conversations including the sync of the participants. Only the participants were synced before. However, if multiple messages need to be stored, the method does not perform any checks so far. In cases of missing references, no inserts will be performed.

ChannelDatabaseManager: Performed the described bug related changes in StoreAnnouncement and BulkInsertOfAnnouncements.
GroupDatabaseManager: Bug related changes in StoreConversationMessage and BulkInsertConversationMessages.
Method IsConversationStored added.

Minor changes in the error codes.

InGroupDetailsViewModel: Changed the implementation of the method LoadGroupFromTemporaryCacheAsync. The problem was that the normal screen with the "join group" command was displayed when clicking on a group in the group search, even if the group was already contained in the local datasets. This happend when the group was locally stored, but the participant has been removed from the group by the admin. Changed this behavior to show the group details view with the "delete locally" option.

Added view updates in response to errors in BallotDetailsViewModel, ConversationDetailsViewModel and GroupDetailsViewModel.

Minor change in group sorting in Homescreen view model.

Added text for "delete locally" command in cases the user has been removed from the group by the admin.
Added the "participant removed" icon in white.