Minor changes in SearchChannel and SearchGroup.
Changed the layout of SearchChannel. The search button is now presented in the commandBar instead next to the text box. If the text box gets focus, the order by type and sync buttons are hidden and displayed again after the focus was removed from the text box.

In GroupSearch is no search perfomed if "search by name" is selected and no check box is chosen.