Added scaled versions of group, conversation and ballot icons. Also took language preferences into account, refer to directories in Assets folder.
Replaced status bar strings in ApplicationsSettingsViewModel and of course in corresponding resource files.

Changed implementation of synchronization command/functionality in ChannelDetails view (user perspective). Replaced the two separate commands with a single command and a distinction of cases.
This change also reflects to ChannelDetails.xaml.

Changed determineLastActivePivotItem and LoadState implementation in code behind of GroupDetails.

Changed some labels in resource files, especially command labels in the views for the moderators.